extract `sign` from test/foreign into test/helpers and improve it
